# api/dist/src/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "schemes": ["http"],
    "produces": ["application/json"],
    "paths": {
        "/news": {
            "get": {
                "tags": ["Manage Articles!"],
                "summary": "This method provides the news of a specified year",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "content": {
                    "application/json": {}
                },
                "parameters": [
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Specifies the query string",
                        "required": true,
                        "type": "number"
                    }
                ]
            },
            "post": {
                "tags": ["Manage Articles!"],
                "summary": "Add a new article to the database",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "News",
                        "description": "Article Object to add",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "required": [
                                "userName"
                            ],
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "body": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "authors": {
                                    "type": "list"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/news/{id}": {
            "put": {
                "tags": ["Manage Articles!"],
                "summary": "Change an existing article",
                "description": "",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id"
                    },
                    {
                        "in": "body",
                        "name": "News",
                        "description": "Article Object replacer",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "title": {
                                    "type": "string"
                                },
                                "body": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "authors": {
                                    "type": "list"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            },
            "delete": {
                "tags": ["Manage Articles!"],
                "summary": "Delete an article!",
                "description": "Delete x article from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id"
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        }
    }
}